requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
finnhub-python==2.4.13
greenlet==1.1.2
idna==3.3
lxml==4.8.0
multitasking==0.0.10
mysql-connector-python==8.0.28
numpy==1.22.3
pandas==1.4.2
protobuf==3.20.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
yfinance==0.1.70
